+++ Sources/tailscale-swift/TailscaleSwift.swift
@@ -24,6 +24,7 @@ struct TailscaleSwift: AsyncParsableCommand {
       NetcheckCommand.self,
       DNSCommand.self,
       CheckForwardingCommand.self,
+      SetCommand.self,
     ]
   )
 }